requirements.txt
# Core
streamlit==1.29.0
PyYAML==6.0
python-dotenv==1.0.1

# LLM / CrewAI
crewai==0.3.0
crewai-tools==0.3.0
langchain-google-genai==0.1.4

# Optional for better performance or utils
requests==2.31.0
